Centralize type exports to types.ts files across all repositories
