+++ src/data/repositories/rbac/index.ts
@@ -1,7 +1,6 @@
 import { assignRole, setUserPermissions } from './mutations'
 import { findInviters, findRole, findRolePermissions, findUserPermissions } from './queries'
 
-export * from './queries'
 export * from './types'
 
 export const rbacRepo = {
