+++ src/data/repositories/user/index.ts
@@ -1,11 +1,6 @@
 import { createUser, deleteUser, updateUser } from './mutations'
 import { findUserByEmail, findUserById, search } from './queries'
 
-export type {
-  SearchParams,
-  SearchResult,
-} from './queries'
-
 export * from './types'
 
 export const userRepo = {
